huggingface-space/requirements.txt
torch
torchvision
gradio
requests
Pillow
